Move signature mismatch into capability module
